chore(api): rename for convention
