+++ src/actions/RouteActions.js
@@ -2,7 +2,7 @@ import {default as Immutable} from "immutable";
 import {default as Rx} from "rx";
 import {FuncSubject} from "rx-react";
 
-import * as ParseAPI from "../api/Parse";
+import * as ParseAPI from "../api/ParseAPI";
 import {default as RouteConstants} from "../constants/RouteConstants";
 
 export default class RouteActions {
